backend/requirements.txt
Django>=4.2
djangorestframework>=3.14.0
python-dotenv>=1.0.0
mysqlclient>=2.1.0
django-cors-headers>=4.0.0
drf-spectacular>=0.26.0
drf-spectacular-sidecar>=0.12.0
gunicorn>=21.2.0
whitenoise>=6.5.0
dj-database-url>=3.0.0
PyMySQL>=1.1.0
requests>=2.31.0
